requirements.txt
# Telegram Bot API
python-telegram-bot==20.7

# Webhook server
aiohttp==3.9.1

# Google Sheets API
google-api-python-client==2.108.0
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1

# QR Code generation
qrcode[pil]==7.4.2
Pillow==10.1.0

# Date/Time handling
python-dateutil==2.8.2
pytz==2023.3

# Environment variables
python-dotenv==1.0.0

# Logging and utilities
colorlog==6.8.0

# Data handling
pandas==2.1.4

# HTTP requests
requests==2.31.0

# JSON handling
ujson==5.8.0 